Fix onboarding on mobile web
